CPM-677: Clean ghost records during uuid migration
